msi-bruisenet/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
tensorboard>=2.10.0
pyyaml>=6.0
tqdm>=4.64.0
thop>=0.1.1
